+++ src/routes/Routes.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { Route, Switch, Redirect } from 'react-router-dom';
 import BrowserProvider from '../util/browser/BrowserProvider';
 
-import RouteValidator from '../auth/routes/RouteValidator';
+import RouteValidator from '../auth/routes/Validator';
 import TodoList from '../pages/Todo/List/TodoList';
 import TodoEdit from '../pages/Todo/Edit/TodoEdit';
 import NotFound from '../pages/Fallback/NotFound';
